refractor of Widgets into new Files and added edit to Shopping List items
