+++ lib/features/shopping_list/presentation/shopping_list_screen.dart
@@ -1,8 +1,7 @@
 import 'package:expandable/expandable.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:mealmaster/db/storage_ingredient.dart';
 import 'package:mealmaster/features/shopping_list/domain/shopping_list_repository.dart';
+import 'package:mealmaster/features/shopping_list/presentation/widgets/shopping_list_grid.dart';
 //TODO: Clean Up and split into seperate Files
 //Missing Features add new Items and Change Items
 
@@ -35,6 +34,8 @@ class _ShoppingListScreenState extends State<ShoppingListScreen> {
                 shoppingListGrid(shoppingList, true, (ShoppingListItem item) {
               repo.removeItemFromShoppingList(item);
               setState(() {});
+            }, () {
+              setState(() {});
             }),
           ),
           ExpandablePanel(
@@ -48,6 +49,8 @@ class _ShoppingListScreenState extends State<ShoppingListScreen> {
                   (ShoppingListItem item) {
                 repo.addRecentItemToShoppingList(item);
                 setState(() {});
+              }, () {
+                setState(() {});
               }),
             ),
           )
@@ -56,84 +59,3 @@ class _ShoppingListScreenState extends State<ShoppingListScreen> {
     );
   }
 }
-
-Widget shoppingListItemCard(final ShoppingListItem item, Function onClick) {
-  return Card(
-    child: ListTile(
-      title: Text(item.name),
-      subtitle: Text("${item.count}"),
-      onTap: () {
-        onClick(item);
-      },
-      onLongPress: () => {},
-    ),
-  );
-}
-
-Widget addShoppingListItemButton(context) {
-  return Card(
-      child: IconButton(
-          onPressed: () {
-            newItemDialog(context);
-          },
-          icon: Icon(Icons.add)));
-}
-
-Widget shoppingListGrid(Future<List<ShoppingListItem>> futureList,
-    bool hasAddButton, Function onClick) {
-  int startIndex = hasAddButton ? 1 : 0;
-  return FutureBuilder(
-      future: futureList,
-      builder: (context, list) {
-        if (list.hasData == false) {
-          return Container();
-        }
-        return GridView.builder(
-            itemCount: startIndex + list.data!.length,
-            gridDelegate:
-                SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 3),
-            itemBuilder: (context, index) {
-              if (index == 0 && hasAddButton) {
-                return addShoppingListItemButton(context);
-              }
-              return shoppingListItemCard(
-                  list.data![index - startIndex], onClick);
-            });
-      });
-}
-
-Future newItemDialog(context) => showDialog(
-    context: context,
-    builder: (context) => AlertDialog(
-          title: Text("Neues Item erstellen"),
-          icon: Icon(Icons.add),
-          content: SizedBox(
-            height: 200,
-            child: Column(
-              children: [
-                TextField(
-                  decoration: InputDecoration(
-                      hintText: "Name des Produktes", labelText: "Name"),
-                ),
-                TextField(
-                  decoration: InputDecoration(
-                      hintText: "Bite gib eine Menge an",
-                      labelText: "MengenAngabe"),
-                  inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                )
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-                child: Text("Abbrechen")),
-            TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-                child: Text("Hinzufügen"))
-          ],
-        ));
